requirements.txt
pyyaml
pytest
pytest-flake8
tox
